requirements.txt
absl-py==0.1.10
alabaster==0.7.9
anaconda-client==1.6.0
anaconda-navigator==1.5
anaconda-project==0.4.1
astroid==1.4.9
astropy==1.3
attrs==17.2.0
Automat==0.0.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.4
boto==2.45.0
boto3==1.5.1
botocore==1.8.15
Bottleneck==1.2.0
bz2file==0.98
cffi==1.9.1
chardet==2.3.0
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
conda==4.4.8
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.4
cryptography==1.7.1
cssselect==1.0.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.13.0
datashape==0.5.4
decorator==4.0.11
dill==0.2.5
docutils==0.13.1
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
futures==3.1.1
gensim==3.2.0
get==0.0.39
gevent==1.2.1
greenlet==0.4.11
h5py==2.6.0
HeapDict==1.0.0
html2text==2018.1.9
html5lib==0.9999999
hyperlink==17.3.1
hypothesis==3.23.0
idna==2.2
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.4
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==2.1.3
lazy-object-proxy==1.2.2
llvmlite==0.15.0
locket==0.2.0
lxml==3.7.2
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.9
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.3.1
numba==0.30.1
numexpr==2.6.1
numpy==1.14.0
numpydoc==0.6.0
odo==0.5.0
openpyxl==2.4.1
pandas==0.19.2
parsel==1.2.0
partd==0.3.7
pathlib2==2.2.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ply==3.9
post==0.0.26
prompt-toolkit==1.0.9
protobuf==3.5.1
psutil==5.0.1
ptyprocess==0.5.1
public==0.0.65
py==1.4.32
pyasn1==0.1.9
pyasn1-modules==0.0.9
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
pyflakes==1.5.0
Pygments==2.1.3
pylint==1.6.4
pymongo==3.6.0
Pympler==0.5
pyOpenSSL==16.2.0
pyparsing==2.1.4
pyspark==2.1.0+hadoop2.7
pytest==3.0.5
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
query-string==0.0.28
queuelib==1.4.2
readability-lxml==0.6.2
redis==2.10.5
request==0.0.26
requests==2.12.4
rope-py3k==0.9.4.post1
s3transfer==0.1.12
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.4.0
seaborn==0.7.1
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.5
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
spyder==3.1.2
SQLAlchemy==1.1.5
statsmodels==0.6.1
sympy==1.0
tables==3.3.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
terminado==0.6
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
Twisted==17.5.0
unicodecsv==0.14.1
w3lib==1.18.0
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
wrapt==1.10.8
xgboost==0.7.post3
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
zope.interface==4.4.2
